Implementation the swagger documentation to other routes, more validations and adjustments

+++ techTest/service/category.js
@@ -1,6 +1,16 @@
 var repository = require('../repository/category');
 var productRepository = require('../repository/product');
 
+function IsValidParameter(object) {
+    if (!object.hasOwnProperty('name') ||
+        !object.hasOwnProperty('description') ||
+        !object.hasOwnProperty('status') ||
+        !object.hasOwnProperty('regDate'))
+        return false;
+
+    return true;
+}
+
 async function GetAll(callback) {
     return await repository.GetAll().then(categories => {
         if (Array.isArray(categories) && categories.length === 0)
@@ -14,45 +24,57 @@ async function GetAll(callback) {
 }
 
 async function GetById(id, callback) {
-    return await repository.GetById(id).then(category => {
-        if (!category)
-            callback(404, 'Category not found', null);
-        else
-            callback(200, 'Success', category);
-    })
+    if (id && id > 0 && !isNaN(id)) {
+        return await repository.GetById(id).then(category => {
+            if (!category)
+                callback(404, 'Category not found', null);
+            else
+                callback(200, 'Success', category);
+        })
         .catch(err => {
             callback(500, 'Error', JSON.stringify(err));
         });
+    } else {
+        callback(403, 'Invalid parameter', null);
+    }
 }
 
 async function Create(category, callback) {
-    return await repository.Create(category).then(inserted => {
-        if (inserted)
-            callback(201, 'Success', null);
-        else
-            callback(500, 'Error', null);
-    })
-        .catch(err => {
-            callback(500, 'Error', JSON.stringify(err));
-        });
-}
-
-async function Delete(id, callback) {
-    let product = await productRepository.GetByCategory(id);
-    
-    if (product.length === 0) {
-        return await repository.Delete(id).then(deleted => {
-            if (deleted)
-                callback(200, 'Success', null);
+    if (IsValidParameter(category)){
+        return await repository.Create(category).then(inserted => {
+            if (inserted)
+                callback(201, 'Success', null);
             else
                 callback(500, 'Error', null);
         })
+        .catch(err => {
+            callback(500, 'Error', JSON.stringify(err));
+        });
+    }else{
+        callback(403, 'Invalid parameter', null);
+    }
+}
+
+async function Delete(id, callback) {
+    if (id && id > 0 && !isNaN(id)) {
+        let product = await productRepository.GetByField('CategoryId',id);
+        
+        if (product.length === 0) {
+            return await repository.Delete(id).then(deleted => {
+                if (deleted)
+                    callback(200, 'Success', null);
+                else
+                    callback(500, 'Error', null);
+            })
             .catch(err => {
                 callback(500, 'Error', JSON.stringify(err));
             });
-    }
-    else {
-        callback(400, 'Cannot be removed, because exists a product associated', null);
+        }
+        else {
+            callback(400, 'Cannot be removed, because exists a product associated', null);
+        }
+    } else {
+        callback(403, 'Invalid parameter', null);
     }
 }
 
@@ -63,9 +85,9 @@ async function Update(id, field, value, callback) {
         else
             callback(500, 'Error', null);
     })
-        .catch(err => {
-            callback(500, 'Error', JSON.stringify(err));
-        });
+    .catch(err => {
+        callback(500, 'Error', JSON.stringify(err));
+    });
 }
 
 module.exports = { GetAll, GetById, Create, Delete, Update };
